Handle event from the leave section
Adds event handlers to the rejoin and return_to_home buttons.
the rejoin button onclick takes the user to the meeting setup once again
and the return_to_home button returns the user to the /apps route if the
user is loggedin otherwise it takes the user to the / route.